Split Sample Track classes into separate files
