+++ include/SampleTCOView.h
@@ -1,0 +1,65 @@
+/*
+ * SampleTCOView.h
+ *
+ * Copyright (c) 2005-2014 Tobias Doerffel <tobydox/at/users.sourceforge.net>
+ *
+ * This file is part of LMMS - https://lmms.io
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation; either
+ * version 2 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program (see COPYING); if not, write to the
+ * Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA 02110-1301 USA.
+ *
+ */
+
+#ifndef SAMPLE_TCO_VIEW_H
+#define SAMPLE_TCO_VIEW_H
+
+#include "SampleTCOView.h"
+
+#include "SampleTCO.h"
+#include "TrackContentObjectView.h"
+
+class SampleTCOView : public TrackContentObjectView
+{
+	Q_OBJECT
+
+public:
+	SampleTCOView( SampleTCO * _tco, TrackView * _tv );
+	virtual ~SampleTCOView() = default;
+
+public slots:
+	void updateSample();
+	void reverseSample();
+
+
+
+protected:
+	void contextMenuEvent( QContextMenuEvent * _cme ) override;
+	void mousePressEvent( QMouseEvent * _me ) override;
+	void mouseReleaseEvent( QMouseEvent * _me ) override;
+	void dragEnterEvent( QDragEnterEvent * _dee ) override;
+	void dropEvent( QDropEvent * _de ) override;
+	void mouseDoubleClickEvent( QMouseEvent * ) override;
+	void paintEvent( QPaintEvent * ) override;
+
+
+private:
+	SampleTCO * m_tco;
+	QPixmap m_paintPixmap;
+	bool splitTCO( const TimePos pos ) override;
+} ;
+
+
+
+#endif
