+++ include/SampleTrack.h
@@ -25,116 +25,14 @@
 #ifndef SAMPLE_TRACK_H
 #define SAMPLE_TRACK_H
 
-#include <QDialog>
 #include <QLayout>
 
 #include "AudioPort.h"
 #include "FadeButton.h"
 #include "FxMixer.h"
-#include "FxLineLcdSpinBox.h"
+#include "SampleTCO.h"
+#include "SampleTrackView.h"
 #include "Track.h"
-#include "TrackContentObjectView.h"
-#include "TrackView.h"
-
-class EffectRackView;
-class Knob;
-class SampleBuffer;
-class SampleTrackWindow;
-class TrackLabelButton;
-class QLineEdit;
-
-
-class SampleTCO : public TrackContentObject
-{
-	Q_OBJECT
-	mapPropertyFromModel(bool,isRecord,setRecord,m_recordModel);
-public:
-	SampleTCO( Track * _track );
-	SampleTCO( const SampleTCO& orig );
-	virtual ~SampleTCO();
-
-	SampleTCO& operator=( const SampleTCO& that ) = delete;
-
-	void changeLength( const TimePos & _length ) override;
-	const QString & sampleFile() const;
-
-	void saveSettings( QDomDocument & _doc, QDomElement & _parent ) override;
-	void loadSettings( const QDomElement & _this ) override;
-	inline QString nodeName() const override
-	{
-		return "sampletco";
-	}
-
-	SampleBuffer* sampleBuffer()
-	{
-		return m_sampleBuffer;
-	}
-
-	TimePos sampleLength() const;
-	void setSampleStartFrame( f_cnt_t startFrame );
-	void setSamplePlayLength( f_cnt_t length );
-	TrackContentObjectView * createView( TrackView * _tv ) override;
-
-
-	bool isPlaying() const;
-	void setIsPlaying(bool isPlaying);
-
-public slots:
-	void setSampleBuffer( SampleBuffer* sb );
-	void setSampleFile( const QString & _sf );
-	void updateLength();
-	void toggleRecord();
-	void playbackPositionChanged();
-	void updateTrackTcos();
-
-
-private:
-	SampleBuffer* m_sampleBuffer;
-	BoolModel m_recordModel;
-	bool m_isPlaying;
-
-	friend class SampleTCOView;
-
-
-signals:
-	void sampleChanged();
-	void wasReversed();
-
-} ;
-
-
-
-class SampleTCOView : public TrackContentObjectView
-{
-	Q_OBJECT
-
-public:
-	SampleTCOView( SampleTCO * _tco, TrackView * _tv );
-	virtual ~SampleTCOView() = default;
-
-public slots:
-	void updateSample();
-	void reverseSample();
-
-
-
-protected:
-	void contextMenuEvent( QContextMenuEvent * _cme ) override;
-	void mousePressEvent( QMouseEvent * _me ) override;
-	void mouseReleaseEvent( QMouseEvent * _me ) override;
-	void dragEnterEvent( QDragEnterEvent * _dee ) override;
-	void dropEvent( QDropEvent * _de ) override;
-	void mouseDoubleClickEvent( QMouseEvent * ) override;
-	void paintEvent( QPaintEvent * ) override;
-
-
-private:
-	SampleTCO * m_tco;
-	QPixmap m_paintPixmap;
-	bool splitTCO( const TimePos pos ) override;
-} ;
-
-
 
 
 class SampleTrack : public Track
@@ -202,126 +100,6 @@ private:
 
 } ;
 
-
-
-class SampleTrackView : public TrackView
-{
-	Q_OBJECT
-public:
-	SampleTrackView( SampleTrack* Track, TrackContainerView* tcv );
-	virtual ~SampleTrackView();
-
-	SampleTrackWindow * getSampleTrackWindow()
-	{
-		return m_window;
-	}
-
-	SampleTrack * model()
-	{
-		return castModel<SampleTrack>();
-	}
-
-	const SampleTrack * model() const
-	{
-		return castModel<SampleTrack>();
-	}
-
-
-	QMenu * createFxMenu( QString title, QString newFxLabel ) override;
-
-
-public slots:
-	void showEffects();
-	void updateIndicator();
-
-
-protected:
-	void modelChanged() override;
-	QString nodeName() const override
-	{
-		return "SampleTrackView";
-	}
-
-	void dragEnterEvent(QDragEnterEvent *dee) override;
-	void dropEvent(QDropEvent *de) override;
-
-private slots:
-	void assignFxLine( int channelIndex );
-	void createFxLine();
-
-
-private:
-	SampleTrackWindow * m_window;
-	Knob * m_volumeKnob;
-	Knob * m_panningKnob;
-	FadeButton * m_activityIndicator;
-
-	TrackLabelButton * m_tlb;
-
-	FadeButton * getActivityIndicator() override
-	{
-		return m_activityIndicator;
-	}
-
-	friend class SampleTrackWindow;
-
-} ;
-
-
-
-class SampleTrackWindow : public QWidget, public ModelView, public SerializingObjectHook
-{
-	Q_OBJECT
-public:
-	SampleTrackWindow(SampleTrackView * tv);
-	virtual ~SampleTrackWindow();
-
-	SampleTrack * model()
-	{
-		return castModel<SampleTrack>();
-	}
-
-	const SampleTrack * model() const
-	{
-		return castModel<SampleTrack>();
-	}
-
-	void setSampleTrackView(SampleTrackView * tv);
-
-	SampleTrackView *sampleTrackView()
-	{
-		return m_stv;
-	}
-
-
-public slots:
-	void textChanged(const QString & new_name);
-	void toggleVisibility(bool on);
-	void updateName();
-
-
-protected:
-	// capture close-events for toggling sample-track-button
-	void closeEvent(QCloseEvent * ce) override;
-
-	void saveSettings(QDomDocument & doc, QDomElement & element) override;
-	void loadSettings(const QDomElement & element) override;
-
-private:
-	void modelChanged() override;
-
-	SampleTrack * m_track;
-	SampleTrackView * m_stv;
-
-	// widgets on the top of an sample-track-window
-	QLineEdit * m_nameLineEdit;
-	Knob * m_volumeKnob;
-	Knob * m_panningKnob;
-	FxLineLcdSpinBox * m_effectChannelNumber;
-
-	EffectRackView * m_effectRack;
-
-} ;
 
 
 #endif
