+++ src/core/CMakeLists.txt
@@ -66,6 +66,7 @@ set(LMMS_SRCS
 	core/SampleBuffer.cpp
 	core/SamplePlayHandle.cpp
 	core/SampleRecordHandle.cpp
+	core/SampleTCO.cpp
 	core/SerializingObject.cpp
 	core/Song.cpp
 	core/TempoSyncKnobModel.cpp
