+++ src/core/SamplePlayHandle.cpp
@@ -28,7 +28,7 @@
 #include "Engine.h"
 #include "InstrumentTrack.h"
 #include "Mixer.h"
-#include "SampleTrack.h"
+#include "SampleTCO.h"
